Missed that one

+++ bundles/net.tourbook.cloud/src/net/tourbook/cloud/suunto/SuuntoTokensRetrievalHandler.java
@@ -91,7 +91,7 @@ public class SuuntoTokensRetrievalHandler extends TokensRetrievalHandler {
    public static boolean getValidTokens() {
 
       if (!OAuth2Utils.isAccessTokenExpired(
-            _prefStore.getLong(Preferences.getSuuntoAccessTokenIssueDateTime_Active_Person_String()) + _prefStore.getInt(
+            _prefStore.getLong(Preferences.getSuuntoAccessTokenIssueDateTime_Active_Person_String()) + _prefStore.getLong(
                   Preferences.getSuuntoAccessTokenExpiresIn_Active_Person_String()) * 1000)) {
          return true;
       }
